requirements.txt
fastapi
uvicorn
snowflake-connector-python
pydantic